// frontend/src/app/models/stavka-korpe.ts
export default class StavkaKorpe {
    sifra: string = "";
    nazivProizvoda: string = "";

    stamparijaId: string = "";
    nazivStamparije: string = "";
    grad: string = "";

    boja: string = "";
    uslugaId: string = "";
    tipStampe: string = "";

    jedinicnaCena: number = 0;
    dodatnaCenaPoKomadu: number = 0;
    kolicina: number = 1;
    ukupnaCena: number = 0;

    tekst: string = "";
    slikaZaStampu: string = "";
}